feat(spec-E #601) Phase 3 commit 1: Decision E split — extract list-pagination
Per architect plan #606 r3 §3 Decision E + §4 module map: extract the
`list` + cursor-pagination cluster out of `conversation.service.ts`
into the new sibling module
`packages/server/src/task/services/conversation/list-pagination.ts`.

Pure-function module — no service class state. Service class keeps
the public `list(...)` method as a thin delegate that wires
`{ db, previewCache }` into `listConversations({deps, input})`. Public
service API unchanged.

Symbols moved:
- `list` (kept as delegate)
- `parseListCursor`
- `queryConversationListRows`
- `archivedListFilter`
- `cursorListFilter`
- `conversationSummariesFromRows`
- `attachSummaryParticipants`
- `participantRefsByConversation`
- `nextConversationListCursor`

`listConversations` is the only exported function (with `Effect.withSpan`
for tracing); the other helpers are module-internal `const` arrows.

Circular-dependency check: new module reads `db` + `previewCache`
only; never `connections` / `participants` / `resolveContactPolicy`
(per plan §3 Decision E rationale). No imports from `app/capabilities/`.

File size after split:
- `conversation.service.ts` → 1045 lines (5-line cushion under
  `max-lines: 1050` lint cap; Phase 3 cutover edits in commits 2-4
  consume the cushion).
- `list-pagination.ts` → 198 lines.
