+++ packages/server/src/task/services/conversation.service.ts
@@ -30,6 +30,7 @@ import {
   transaction,
 } from "../../db/effect-kysely-toolkit.js";
 import { assertConversationAdminAuthority } from "./conversation-admin-authority.js";
+import { listConversations } from "./conversation/list-pagination.js";
 import type {
   AddParticipantOptions,
   ContactEdgeInput,
@@ -38,7 +39,6 @@ import type {
   ConversationColumns,
   CreateConversationOptions,
   CreatorContactPolicyInput,
-  ListRow,
   ParticipantAddedBroadcast,
   ParticipantInsertResult,
   ParticipantRemovedBroadcast,
@@ -316,143 +316,10 @@ export class ConversationService {
     { conversations: ConversationSummary[]; cursor?: string },
     InvalidParamsError
   > {
-    return catchSqlErrorAsDefect(
-      Effect.gen(this, function* () {
-        const cursorParam = yield* this.parseListCursor(cursor);
-        const rows = yield* this.queryConversationListRows({
-          agentId,
-          limit,
-          cursorParam,
-          archived,
-        });
-        const hasMore = rows.length > limit;
-        const resultRows = hasMore ? rows.slice(0, limit) : rows;
-        const conversations = this.conversationSummariesFromRows(resultRows);
-        yield* this.attachSummaryParticipants(conversations);
-        return {
-          conversations,
-          cursor: this.nextConversationListCursor(hasMore, resultRows),
-        };
-      }),
+    return listConversations(
+      { db: this.db, previewCache: this.previewCache },
+      { agentId, limit, cursor, archived },
     );
-  }
-
-  private parseListCursor(
-    cursor: string | undefined,
-  ): Effect.Effect<string | null, InvalidParamsError> {
-    if (cursor == null) return Effect.succeed(null);
-    const parsed = new Date(cursor);
-    if (Number.isNaN(parsed.getTime()) || parsed.toISOString() !== cursor) {
-      return Effect.fail(
-        new InvalidParamsError({
-          message: "Cursor must be an ISO-8601 timestamp",
-        }),
-      );
-    }
-    return Effect.succeed(cursor);
-  }
-
-  private queryConversationListRows(input: {
-    readonly agentId: AgentId;
-    readonly limit: number;
-    readonly cursorParam: string | null;
-    readonly archived: ConversationArchiveFilter;
-  }): Effect.Effect<ReadonlyArray<ListRow>, SqlError> {
-    return rawQuery(
-      this.db,
-      sql<ListRow>`
-        SELECT c.id, c.type, c.name, c.updated_at,
-               m.parts_encrypted IS NOT NULL as has_last_message,
-               m.created_at as last_message_at,
-               COALESCE(
-                 (SELECT COUNT(*) FROM messages m2
-                  WHERE m2.conversation_id = c.id
-                  AND m2.seq > cp.last_read_seq
-                  AND m2.is_deleted = false), 0
-               )::int as unread_count
-        FROM conversation_participants cp
-        JOIN conversations c ON c.id = cp.conversation_id
-        LEFT JOIN LATERAL (
-          SELECT parts_encrypted, created_at, seq FROM messages
-          WHERE conversation_id = c.id AND is_deleted = false
-          ORDER BY seq DESC LIMIT 1
-        ) m ON true
-        WHERE cp.agent_id = ${input.agentId}
-          ${this.archivedListFilter(input.archived)}
-          ${this.cursorListFilter(input.cursorParam)}
-        ORDER BY COALESCE(m.created_at, c.updated_at) DESC
-        LIMIT ${input.limit + 1}
-      `,
-    );
-  }
-
-  private archivedListFilter(archived: ConversationArchiveFilter) {
-    switch (archived) {
-      case "only":
-        return sql`AND c.archived_at IS NOT NULL`;
-      case "include":
-        return sql``;
-      case "exclude":
-        return sql`AND c.archived_at IS NULL`;
-    }
-  }
-
-  private cursorListFilter(cursorParam: string | null) {
-    if (cursorParam === null) return sql``;
-    return sql`AND c.updated_at < ${cursorParam}`;
-  }
-
-  private conversationSummariesFromRows(
-    rows: ReadonlyArray<ListRow>,
-  ): ConversationSummary[] {
-    return rows.map((row) => ({
-      id: row.id,
-      type: row.type,
-      name: row.name ?? undefined,
-      lastMessagePreview: this.previewCache.get(row.id),
-      lastMessageTimestamp: row.last_message_at?.toISOString(),
-      unreadCount: row.unread_count,
-    }));
-  }
-
-  private attachSummaryParticipants(
-    conversations: ConversationSummary[],
-  ): Effect.Effect<void, SqlError> {
-    if (conversations.length === 0) return Effect.void;
-    return Effect.gen(this, function* () {
-      const convIds = conversations.map((conversation) => conversation.id);
-      const rows = yield* this.db
-        .selectFrom("conversation_participants")
-        .select(["conversation_id", "agent_id"])
-        .where("conversation_id", "in", convIds);
-      const partsByConv = this.participantRefsByConversation(rows);
-      for (const conversation of conversations) {
-        conversation.participants = partsByConv.get(conversation.id) ?? [];
-      }
-    });
-  }
-
-  private participantRefsByConversation(
-    rows: ReadonlyArray<{ conversation_id: ConversationId; agent_id: AgentId }>,
-  ): Map<ConversationId, Array<{ type: "agent"; id: AgentId }>> {
-    const partsByConv = new Map<
-      ConversationId,
-      Array<{ type: "agent"; id: AgentId }>
-    >();
-    for (const row of rows) {
-      const participants = partsByConv.get(row.conversation_id) ?? [];
-      participants.push({ type: "agent", id: row.agent_id });
-      partsByConv.set(row.conversation_id, participants);
-    }
-    return partsByConv;
-  }
-
-  private nextConversationListCursor(
-    hasMore: boolean,
-    rows: ReadonlyArray<ListRow>,
-  ): string | undefined {
-    if (!hasMore) return undefined;
-    return rows[rows.length - 1]?.updated_at.toISOString();
   }
 
   get(
